+++ src/features/about/AboutBEPopup.cpp
@@ -1,15 +1,11 @@
 #include "AboutBEPopup.hpp"
 #include "ChangelogPopup.hpp"
-#include "SupportPopup.hpp"
 #include <Geode/ui/GeodeUI.hpp>
 #include <Geode/ui/MDTextArea.hpp>
 #include <Geode/binding/ButtonSprite.hpp>
 #include <Geode/utils/web.hpp>
-#include <utils/Pro.hpp>
-
-#ifdef BETTEREDIT_PRO
-#include <pro/features/supporters/SupportersPopup.hpp>
-#endif
+#include <features/supporters/SupportersPopup.hpp>
+#include <features/supporters/ActivateLicensePopup.hpp>
 
 struct Dev {
     const char* name;
@@ -157,9 +153,7 @@ bool AboutBEPopup::setup() {
     menu->setLayout(RowLayout::create()->setGrowCrossAxis(true));
     m_mainLayer->addChildAtPosition(menu, Anchor::Center, ccp(0, -60));
 
-#ifdef BETTEREDIT_PRO
     pro::addAboutPopupStuff(this);
-#endif
 
     // BE links
 
@@ -208,15 +202,24 @@ void AboutBEPopup::onClose(CCObject* sender) {
 }
 
 void AboutBEPopup::onSupport(CCObject*) {
-    SupportPopup::create(false)->show();
-    // openSupportPopup(Mod::get());
+    createQuickPopup(
+        "Support BetterEdit",
+        "The <cp>Supporter Perks</c> for BetterEdit are <co>about to be "
+        "discontinued</c> as my financial situation has gotten better and "
+        "<cj>I don't want to paywall features</c>. If you still want to support the mod "
+        "and donate a thank you, feel free to do so, but be aware that the only "
+        "perk I can offer is your name in the Supporters List!",
+        "OK", "Activate Existing License",
+        380,
+        [](auto, bool btn2) {
+            if (btn2) {
+                ActivateLicensePopup::tryShowIfLoggedIn();
+            }
+        }
+    );
 }
 void AboutBEPopup::onSupporters(CCObject*) {
-#ifdef BETTEREDIT_PRO
     pro::SupportersPopup::create()->show();
-#else
-    FLAlertLayer::create("BE Supporters", "Coming soon ;)", "OK")->show();
-#endif
 }
 
 void AboutBEPopup::onSuggestFeature(CCObject*) {