+++ src/utils/Pro.hpp
@@ -1,9 +1,0 @@
-#pragma once
-
-#ifdef BETTEREDIT_PRO
-#include <pro/Pro.hpp>
-#endif
-
-namespace be {
-    bool isProUIEnabled();
-}